Add NavBar special case for external wiki link
Per https://github.com/ReactTraining/react-router/issues/1147 , <Link/>
is a router-aware anchor and therefore doesn't handle full URLs. The old
NavBar created Menu.Items (which are a Semantic UI React component)
which were augmented as Links (which are actually HashLinks due to the
import statement). This change adds a ternary expression to change the
augmentation of a NavBar link to a raw <a/> tag if it contains a truthy
"external" property.